Move common code to processing_threads.cpp
