+++ tp02/build.sh
@@ -4,5 +4,5 @@ mkdir -p bin
 
 part=part2
 
-g++ -Wall -O2 -D_REENTRANT "src/$part.cpp" src/preprocess.cpp src/utils.cpp -o "bin/$part" -lpthread -lrt
+g++ -Wall -O2 -D_REENTRANT "src/$part.cpp" src/preprocess.cpp src/processing_threads.cpp src/utils.cpp -o "bin/$part" -lpthread -lrt
 sudo "./bin/$part"
